app/requirements.txt
streamlit
pandas
requests
pydantic
pyyaml
altair
